refactor(rlm): share prompt cache usage helper
